FEAT: 챔피언 ID Long으로 변경

+++ src/main/java/com/clonelol/summoner/api/summonerapi/dto/property/infoProperty/Participant.java
@@ -10,7 +10,7 @@ public class Participant {
     private int bountyLevel;
     private int champExperience;
     private int champLevel;
-    private int championId;
+    private Long championId;
     private String championName;
 
     /*
